switch to Astro; add Discord link
initialized a new Astro site, got it looking good, then copied over the
old `.git` folder, so this commit looks big and messy.

Only actual change to the home page is the addition of a link to our
Discord server.

But also no JS on the deployed site! Because Astro.